The trans_pv_sv program takes snapshot files [S,T,U,V,SSH]."timestepcount".[data,meta] and creates the 'assim_model_state_ud' file appropriate to be fed to DART. The time tag IN the file is derived from the middle portion of the snapshot filename and knowledge of the model deltaTmom, startDate_1, startDate_2 namelist values. The "timestepcount" is user input (read from stdin).
The trans_sv_pv program takes DART-produced 'assim_model_state_ic' files
that have both the valid time and the target (advance_to_) time and produces:
  S.yyyymndd.hhmmss.[data,meta]
  T.yyyymndd.hhmmss.[data,meta]
  U.yyyymndd.hhmmss.[data,meta]
  V.yyyymndd.hhmmss.[data,meta]
SSH.yyyymndd.hhmmss.[data,meta] files as well as a new 'data.cal.new'
namelist file with the updated startDate_1 and startDate_2 values. There
is still work to be done to update the PARM03 namelist variable 'endTime'
with the value written to stdout, and how we update the PARM05 namelist
with the values of the new initial files.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@3308 acad1b26-7d42-44b9-9cec-aac8293558f5