Added support for variables with UNDEFINED vertical levels. They are now treated as a 'one-level' variable - meaning they are all glommed together. The matlab scripts also support UNDEFINED.
And, Just For Nancy, I fixed the logic of the program so that now
if you have JUST ONE observation sequence file in another directory,
obs_diag will process it and terminate normally rather than die if
the directory node does not have an 'obvious' way to increment it.
i.e.:
   obs_sequence_name = '/Users/thoar/temp/obs_seq.final',
works.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@3457 acad1b26-7d42-44b9-9cec-aac8293558f5