Major update to the radar observation module, along with matching documentation and minor updates to the obs_kind and wrf model_mod files.  The following differences are not backward compatible:
- the wrf model_mod used to stop if model_interpolate() was called with an
unknown kind.  it now sets an error istatus code and returns.

- the radar namelist has changed significantly.  see the documentation html
file for details of the new namelist values.

- the interactive create routines now take an azimuth and elevation for the
radar beam and then do the trig to store 3 values in the file.  older versions
of this code required the user to do the math themselves.

- some constants which depend on the particular microphysics scheme being
used have changed.  the defaults match the default microphysics scheme in
wrf, but users should verify the values are what they expect.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@3657 acad1b26-7d42-44b9-9cec-aac8293558f5