Major change in this commit is the addition of an observations directory. Eventually all the programs which convert observations of various types into DART obs_seq files will be under this directory. For starters, there are 3: GPS, Radar, and MADIS converters.  Each has documentation.
Also, committing the obs_sequence_tool which is the successor to the
merge_sequence utility.  It has a lot more options; see the docs.
Rename restart_file_utility to restart_file_tool to be consistent.

Minor update to the intel mac template file.

Updates to the obs_sequence_mod required by the obs_sequence_tool;
mostly additional functions, but also removed the fixed 129 char
length specification on many of the incoming args to the subroutines
here.  No need to force the length; makes it easier on the callers.
Also removed unused constants from the types_mod use statement.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@3688 acad1b26-7d42-44b9-9cec-aac8293558f5