Everything compiles on coral with Intel 10.1.
The pop_to_dart code runs and is to be thoroughly checked by the
matlab script Check_pop_to_dart.m. At present, I think matlab
unwraps the N-D variables in opposite order to the way they are
handled in the Fortran code.

The model_mod:static_init_model() routine checks out, but the
stubs for orthogonality are just that - stubs.

The communication of how often it is possible to stop the model
and when to generate the restart file is not finished.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@3945 acad1b26-7d42-44b9-9cec-aac8293558f5