Glen and I added code to allow reflectivity kinds to be interpolated, and allowed them to be gracefully failed if no reflectivity fields are in the state vector.  the default obs_kind mod for radar first tries to interpolate reflectivity directly, and then backs off and tries to compute it from the other fields if necessary.
Also removed the remaining dependance on nmoist and surf_obs
from the namelist. Renamed QGRAUPEL to QGRAUP in code to match
the actual name of the field in the wrf netcdf files.
And finally, added a lot more checks to be sure the starting 
index is valid before using it to access into the state vector.


git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@3992 acad1b26-7d42-44b9-9cec-aac8293558f5