The obs_diag program now has the following 'copies': Nposs Nused NbigQC NbadIZ NbadUV NbadLV rmse bias spread totalspread NbadDARTQC observation ens_mean N_DARTqc_0 N_DARTqc_1 N_DARTqc_2 N_DARTqc_3 N_DARTqc_4 N_DARTqc_5 N_DARTqc_6 N_DARTqc_7
which can be plotted with the usual diagnostics/matlab routines.
So - now it is possible to not only plot the bias as a function of height,
you can now plot the (mean) observation value as a function of height.
Additionally, obs_diag and the associated plotting functions
also support observations on model levels (curious obs).



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@4168 acad1b26-7d42-44b9-9cec-aac8293558f5