When skipping an observation for any reason the code has to goto 200 (the bottom of the loop) and not directly back to 10 (the top of the loop) or it might read past the end of the input file and crash in the prepbufr libs.
Also wrapped 2 lines that were longer than 72 chars. these
are fixed-format files and without extra flags to extend the
input line length, with at least one compiler characters 
beyond column 72 were silently ignored. in both cases it 
was a debug print message that isn't called unless you 
turn a local debugging variable on.  what was annoying was 
that what was being ignored was the value of interest it 
was supposet to be printing.


git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@4346 acad1b26-7d42-44b9-9cec-aac8293558f5