Ryan Torn's changes to add additional metadata to the netCDF diagnostic files when writing the state data as a single vector.  There is no change to the code in the more common case of breaking the state vector into the consitituent variables, U,V,T, etc and writing them in separate netCDF variables.  The new additional metadata records which variables are in the state vector, what order they are in, and what size they are. This allows a postprocessing program to extract the various fields at a later point.
Also included in this commit is a minor rearrangment of where the module global variables
are declared, to collect the 5 deprecated namelist items together with a comment about them
having no impact on the code.  This has been true for a while now, but it wasn't as
clear as it could have been that setting those items has no impact on the code.  Removing
them from the namelist is a non-backwards-compatible change (users will get a fatal error
at runtime if they remain in the users namelist), but they should be removed as soon as
we freeze the code and make a release.


git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@4369 acad1b26-7d42-44b9-9cec-aac8293558f5