This advances the LANL/POP 2_0_1 version of POP ... but still cannot be used for assimilation until the POP code is modified to do a forward euler timestep.  This compiles and runs on coral (SLES10) with the ifort (IFORT) 10.1 20090203 compiler with the following flags: FFLAGS = -O0 -fpe0 -vec-report0 -assume byterecl
Coral is an Intel-based machine - so all binary files were little-endian.
I like to append a ".le" suffix on those files.

It was checked in the gx3v5 configuration.

For coral with the openmpi framework, it is necessary to
specify input.nml:&mpi_utilities_nml:reverse_task_layout = .true., 
For bluefire ... it must be .false. (the default).



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@4398 acad1b26-7d42-44b9-9cec-aac8293558f5