Change the routine which returns the location of the state vector points. The original version only had the mass grid to work from, and so computed where the staggered points for the U and V grids should be.  However with a deformed grid these were not exactly in the right places.  The code was recently changed to read in the U and V staggered grids for writing to the output diagnostic files, so those grids are now available for use by this routine.  For the U and V staggers, look up the point locations directly in the arrays instead of doing a computation.
note that this code is now more accurate, and the expected differences are
slight, but it will NOT be bitwise reproducible with previous versions.  
in simple tests over the CONUS domain i saw differences in the output 
state space diagnostics of order 10e-4 in the U and V field values.


git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@4495 acad1b26-7d42-44b9-9cec-aac8293558f5