Made the "use outlier observations in the rank histogram" a namelist option. The default is to NOT use those observations in the rank histogram. You MUST specify 'outliers_in_histogram = .true.' if you want them. The netcdf file also contains the appropriate list of DART QC values that went into the rank_histogram, as well as a logical that reiterates the value of the namelist option.
The 'horizontal_wind' "observations" can not get a rank histogram, so I put in logic
to ensure they never get written to the netcdf file. The U,V components do, but the
derived 'speed' does not. There is a log file message (an run-time) that specifies
which variables have NO information in the rank histogram. 



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@4529 acad1b26-7d42-44b9-9cec-aac8293558f5