Removed the transparency of the 'stripe' boxes that denote the definition of the vertical levels. After calling the 'stripes' routine, the background axis and grid can be brought to the foreground by setting the axis 'Layer' property to 'top'. (this is what I tried to achieve by using transparency)
This had several positive effects:

1) using transparency required the use of the OpenGL renderer,
which is buggy and caused the output to be rasterized. Even the .ps files.
Fonts were not rendered as real fonts, it was just ugly.
Because the 'painters' renderer can be used again, the file sizes are 
quite compact. Files went from 11395232 bytes to 27223 bytes! AND they resize
better. Win-Win.

2) the OpenGL renderer had a bug in it that cause spurious label elements
to appear. For the dual-axis labels, I grab the ticks from the ones that are
'nice' and determine the corresponding values for the other value and create
and plot a character string to that effect. The OpenGL version had a bug
in it that left the 'multiplier' floating about. For example, the number of
observations is 10,000 for example. Matlab might plot it as 1.0 x 10^4.
I grab it and actually plot it as 10000 - but the OpenGL renderer left the
stupid x 10^4 string on the plot! Not good.

I have yet to hunt down all the other places this strategy is used.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@4612 acad1b26-7d42-44b9-9cec-aac8293558f5