Several critical fixes for those users who redefine R8 to be R4.
1. Compute variance and covariance in more numerically stable ways
if running with single precision reals.  The previous code was fine
when running with real*8, but the updated code retains more accuracy
when running real*4.  Could have had a noticeable impact on results.

2. Fix minor issues with the Sampling Error Correction; a few of
the constants were incorrect and there was the possibility of an
out-of-range array index being computed.  In general the effects
were minor.

3. Several numerical constants were explicitly marked _r8 to set
the kind.

4. Added an informational message to the dart log if running
with the sampling error correction enabled.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@4869 acad1b26-7d42-44b9-9cec-aac8293558f5