Fix for the long-standing '3d box bug'.  At any latitude except the equator the furthest longitude that is 'maxdist' away by a great circle calculation does not happen at the same latitude, so you can't just add and subtract maxdist to find out the min and max longitude that is still within maxdist of a point.  This is why all the computations here (except for where this bug arises) are done in radians.  The one place where the computations were adding distances along latitude lines was where the code was trying to see if all the locations for this particular task are clustered and so the boxes can be restricted to a subregion of the sphere.  The code now uses great circle distances and computes the latitude at which the longitude is the furthest away.   The code also bypasses some of the computations if it's clear that the boxes will have to cover the entire range of longitudes.
Update to the documentation to include a section on the geometry for
this computation, just because it's fun.

Also minor changes to the formatting of the debug code, and a fix for 
COMPARE_TO_CORRECT. If the optional distance argument is not supplied
this routine returns all locations in nearby boxes, which is a superset of
the actual number of locations within maxdist.  It was doing an exact
compare in all cases, which is too restrictive.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@4900 acad1b26-7d42-44b9-9cec-aac8293558f5