Changes in how we set the mean and stddev when there are failed forward operators.  The older code set the values right before they were written out.  This code sets the values inside the obs loop right when they're computed, and in addition has a section to set the posterior values if any of the posterior forward operators failed, even if the prior qc is bad (the particular case where this is possible is when the prior fails the outlier threshold test, but all the prior forward operators worked).   Also make the same changes to the radar folded version of filter to keep it in sync.
Change the names of OBS_PRIOR_MEAN_... and OBS_PRIOR_VAR_...
to just OBS_MEAN_... and OBS_VAR_... because they are used both
when computing the prior and posterior values.  these are internal
to the filter_main() subroutine and so are not visible outside the
subroutines in which they are used.  just for a bit more clarity.
(there's already a separate flag passed down to the obs diag routine 
to say whether we're computing the prior or posterior.)


git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@4959 acad1b26-7d42-44b9-9cec-aac8293558f5