Model mod write restart file routine working now - was idiotic level-of-indirection error into the progvar() array and nothing to do with the complicated block indexing in the restart files. (i violated the 'look at the simple possibilities first before looking at the most complicated' principle here.)
Also fixed bug in dart_to_gitm where it was writing the time
information in the reverse order (labeled start as end and
vise versa)

added more code to model_mod_check to write and reread
state vectors and test the results for identity.

jumbled the order of the fields in the input.nml file
mostly for testing.  probably should be put back in some
kind of logical order for real cases. 



git-svn-id: https://subversion.ucar.edu/DAReS/DART/branches/gitm@5182 acad1b26-7d42-44b9-9cec-aac8293558f5