match changes in trunk/kodiak:
fix a bug in the termination code - if the end time was exactly
one window before 0Z, it would fail the end test and loop forever.

also add an option (off by default for backwards compatibility)
to allow the script to loop over missing input files.  if some
time periods are missing some types of input files there is now
an option to complain on standard output but continue to run.

also added a couple tests to be sure the required executables
exist and point to where they are built if they aren't found.


git-svn-id: https://subversion.ucar.edu/DAReS/DART/branches/development@5936 acad1b26-7d42-44b9-9cec-aac8293558f5