match changes in trunk/kodiak:
removing an old binary restart file and unused subdirectory.
this version of the bgrid code does not support the reading 
and writing of fms-type restart files without the MPI implementation.  
the grid sizes, levels, etc are specified in a namelist 
and the initial data values are taken from the dart IC files.

git-svn-id: https://subversion.ucar.edu/DAReS/DART/branches/development@5939 acad1b26-7d42-44b9-9cec-aac8293558f5