Moved the test for the outlier_threshold into a separate subroutine at the end of the file.  Added comments about how users can customize this subroutine if they want a different threshold based on location or observation type. There is a new namelist item, enable_special_outlier_code, which can be turned on and off at run time to make it easy to run comparisons between the original simple threshold and any more complex handling added by the user.
Also added a few new lines to the log file - if inflation
damping is on, echo that value.  Echo when calling the
initial and end routines in the model_mod, to help isolate
problems if the code fails.  Also call end_model() for
the first time. It should have been called all along
and never was.  our bad.  Make the qc threshold code
a bit more modular by not referencing a global var.

Identical changes to filter.dopplerfold.f90 to keep
in sync with filter; changes to the default nml and
html pages to document the new namelist item.

All this being said, there should be no changes to
the filter results from this update; there will be
a few new lines in the log files but that's all.


git-svn-id: https://subversion.ucar.edu/DAReS/DART/branches/development@6011 acad1b26-7d42-44b9-9cec-aac8293558f5