These routines support the SQG model and use (only) the snctools toolbox. I have started to replace the calls to deprecated Matlab functions.
The startup.m function is much more graceful and readable.
It will not error out if you are not using DART, for example.

There is a new check in the plot_profile.m function to detect
duplicate 'level' items - which cause a logical conundrum.
The old version of obs_diag.f90 did not check for this - problematic.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/branches/development@6125 acad1b26-7d42-44b9-9cec-aac8293558f5