The model_mod:get_model_time_step() routine now returns the 'model stride', (as opposed to the dynamical timestep of the model). Consequently, the model can be advanced in async=0 mode with no special modifications to the obs_model_mod:adv_1step interface.
A new pair of namelist variables (assimilation_period_[seconds,days])
has been added to allow for control over the model stride.
HOWEVER ... I have not checked to make sure the model stride is a
multiple of the dynamical timestep of the model. As committed to the
repository, the dynamical timestep is 0.0108 * 1000000/30 = 360 seconds
and the model stride is 6 hours - so we're OK for this case.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/branches/development@6149 acad1b26-7d42-44b9-9cec-aac8293558f5