Added a debug print block that summarizes the values of r8,r4, etc. if the print_debug logical is .true.
Also had a bugfix in the next_file() routine. If the user supplies a 
single filename that has a '/' in it - next_file() is supposed to
parse the directory name and increment the portion after the underscore.
dir_0023/obs_seq.final   becomes  dir_0024/obs_seq.final
Problem is ... test_dir/obs_seq.final used to throw an error.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/branches/development@6168 acad1b26-7d42-44b9-9cec-aac8293558f5