assimilate.csh   ... change to use the 'land' observation sequence files. These files have both the Ameriflux and MODIS snow cover fraction observations in them.
perfect_model.csh is enhanced to ensure the casename is correct - needed for
the forward observation operator that reads the CLM history files.

CESM1_1_setup.csh now uses the yellowstone LSF-safe method for error-ing out
and many of the features we use for the models/CESM/shell_scripts/ setup scripts.
This also has a more sensible PE layout for larger ensembles, although it is
still not believed to be anywhere near optimal.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/branches/development@6195 acad1b26-7d42-44b9-9cec-aac8293558f5