Extended GetNCindices.m to be able to read more coordinate dimensions, specifically those used in obs_diag_output.nc.
Compare_netCDF_files is entirely new and will print a summary of the min
and max differences of every numeric variable (regardless of shape) for a
single netCDF file or the difference between a pair of netCDF files.

The CESM_hop_check.m function codifies what we did to ensure the CESM
hop test went well. Hopefully this can be a guide to others.

CompareObsDiag.m is almost like Compare_netCDF_files expect that it
allows one to specify a hyperslab to compare. I had two versions of
obs_diag.f90 that created variables with different numbers of 'copies'
that I needed to compare.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/branches/development@6198 acad1b26-7d42-44b9-9cec-aac8293558f5