fixes for the serious specific types/generic kinds mixup. now in all get_close_obs() calls the base is always passed as a specific type, and the target list is always a list of generic kinds.  where possible i changed the names of things to be accurate.
in assim_tools there is new code to check for either
state vector items with 0 spread, or observations with
0 error variance.  these are handled before the filter_kind
routines are called, so 1) they are all handled in a consistent
way and 2) the specific filter_kind routines don't have to do
error checking (some used to, some didn't - one user reported
a NaN error in filter kind 3 when he had 0 spread in the
ensemble members).  these cases will now all be caught and
handled consistently.  

added a new namelist item to allow disabling of the buffering 
for searches when observations in the same location.  for
things like U, V winds, or T, U, V readings from radiosondes
at mandatory levels where several successive observations are
at identical locations, the buffering can be a large time savings.
but if the model_mod has get_close_obs() code which depends
on looking at the specific type of the current obs, you can't
buffer previous searches at the same location.  the buffering
is on by default, but the user can disable it now.

extensive internal changes in the DEFAULT_obs_kind_mod.F90
to try to separate out the specific types and generic kinds
inside the routines.  next release we may be forced to change
the names of many of the subroutines because most of the ones
with 'kind' in their names are actually returning types.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/branches/development@6200 acad1b26-7d42-44b9-9cec-aac8293558f5