Removed the command-line input mechanism for trans_sv_pv and trans_pv_sv in favor of the namelist-driven mechanism that is used by everyone else. The command-line input was not supported by the gfortran compiler (which precipitated the change).
The model_mod.f90 did not use the 'module_initialized' strategy,
so this is now implemented for all the public interfaces.
The test to check if the netCDF dimensions are as expected
was rewritten to be compiler-agnostic.

The coupler.res file was added as it is needed to by trans_pv_sv.f90



git-svn-id: https://subversion.ucar.edu/DAReS/DART/branches/development@6208 acad1b26-7d42-44b9-9cec-aac8293558f5