one bug fix and some minor formatting/output changes.
the bug fix was in the vertical interpolation in log space;
the equation was missing a critical set of parens.  the choice
of vertical interpolation in linear or log space is now a
namelist option.  the default is log to be consistent with
how wrf does vertial interpolation, but it can be disabled
to be consistent with previous behavior.

i removed several sets of output lines that were duplicated
in the conversion routines.  for both converters there is a
namelist option that allows you to print out the min/max of
each of the fields as they are converted to catch bad values;
or it can be turned off to avoid excessive output if you're
converting 10s to 100s of states at once in parallel.


git-svn-id: https://subversion.ucar.edu/DAReS/DART/branches/development@6242 acad1b26-7d42-44b9-9cec-aac8293558f5