in the xxx_assimilate.csh scripts remove the last reference to the DART directory.  the setup scripts will now copy the sampling error correction file into the CASEROOT dir so these scripts can pick it up from there if it is needed.
remove the deprecated CESM1_1_1_setup.csh script;
rename the other scripts so there is a 'initial' script
which has code to pick up the starting files from
various other places.  the 'continuation' script
assumes there's an existing experiment where the
contents of the 'rest' directory contains everything
you need to begin.

moved code to copy over the final_full.xx file if
needed for the sampling error correction into the
setup scripts.  tried to make the two scripts as
similiar as possible.

added code to the end of the short term archive
script, partially based on alicia's changes.
it does two things - one is that it only leaves
3 restart directories a month in the archive
dir for long term archiving.  if you have to
go back and restart, it will be from the 1st,
the 10th, or the 20th.  if you want to save
more steps, edit the end of the st_archive.sh
and change the test for the day numbers.

one thing that the archive script will do now
is it will also keep the restart files from the 
very last successful step until the next step
has succeeded - both model advance and assimilation.   

so if a step fails the last step can be restarted.
these restart files are not visible to the long term
archiver, so they won't be archived.  they are stored
in the archive directory in a subdir called .sta2
which is hidden unless you use 'ls -al' to see it.
(this is the same strategy the short term archiver
uses to protect partially archived dirs from being
picked up by the long term archiver.)

the 'reset_last_successful_step' script should now
try to find the .sta2 files first before looking
in the 'rest' directory.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@6287 acad1b26-7d42-44b9-9cec-aac8293558f5