Removing 'rat_cri' from all obs_diag namelists. rat_cri has been deprecated for some time now and should have been removed from the namlists a long time ago. rat_cri _was_ important before the DART QC mechanism was supported. rat_cri _used_ to be the only way obs_diag could discriminate good observations from bad ones. obs_diag V 3180 2007-12-19 22:30:32 -0700 (Wed, 19 Dec 2007) was the first version of obs_diag to deprecate 'rat_cri'.
The obs_diag code still defaults to a gigantic number for rat_cri.
No meaningful processing happens based on its value. The only thing
that happens is that the NbadIZ count (preserved in the output netCDF 
file) will always be zero. This is not used by any program, 
so it is no loss.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@6469 acad1b26-7d42-44b9-9cec-aac8293558f5