have the rank histogram filter call a local norm_inv() routine instead of the stats toolbox norminv().
have the oned_model call a local kurt() routine
instead of the stats toolbox kurtosis().

have plot_gaussian call a local norm_pdf() routine
instead of the stats toolbox normpdf().

have the run_lorenz_96 code call different increment
routines based on the menu button for EAKF, EnKF, or RHF.

add a new norm_inv() routine which computes the
inverse CDF for a normal distribution, either for
a single scalar or for an array.  based on the
fortran code found in assim_tools_mod.f90.

with this update, DART_LAB should be completely
independent of the statistics toolbox.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@6831 acad1b26-7d42-44b9-9cec-aac8293558f5