Increased the filename lengths to 256, error message strings to 512. Refactored error messages to make a single call to error_handler() when possible.
Fully specified the default values in the example namelist,
and reordered them to group all 'copy' operations together,
all 'qc' operations together.

Changed documentation to match.

The default 'new_qc_data' had a mismatch in the documentation/example and the
program. The program specified a default of MISSING_R8, the namelist
and documentation had 0.0. Nancy and I decided to change the documentation.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@7145 acad1b26-7d42-44b9-9cec-aac8293558f5